Remove old TaskTest base class.
- Rename pants_test/tasks/test_base.py to task_test_base.py.
- Move TaskTestBase into that file.
- Fix all BUILD file and import references.
