+++ contrib/scrooge/tests/python/pants_test/contrib/scrooge/tasks/BUILD
@@ -19,7 +19,7 @@ python_tests(
     'src/python/pants/backend/jvm/targets:java',
     'src/python/pants/base:build_file_aliases',
     'src/python/pants/base:exceptions',
-    'tests/python/pants_test:task_test_base',
+    'tests/python/pants_test/tasks:task_test_base',
   ],
 )
