+++ tests/python/pants_test/android/BUILD
@@ -42,7 +42,7 @@ python_library(
     'src/python/pants/backend/android/tasks:all',
     'src/python/pants/util:contextutil',
     'src/python/pants/util:dirutil',
-    'tests/python/pants_test:task_test_base'
+    'tests/python/pants_test/tasks:task_test_base'
   ]
 )
