+++ tests/python/pants_test/backend/jvm/tasks/BUILD
@@ -33,7 +33,7 @@ python_tests(
     'src/python/pants/backend/jvm/tasks:ivy_imports',
     'src/python/pants/base:build_file_aliases',
     'src/python/pants/util:contextutil',
-    'tests/python/pants_test:task_test_base'
+    'tests/python/pants_test/tasks:task_test_base'
   ]
 )
 
@@ -82,7 +82,6 @@ python_tests(
     'src/python/pants/backend/jvm/targets:jvm',
     'src/python/pants/backend/jvm/tasks:unpack_jars',
     'src/python/pants/base:target',
-    'tests/python/pants_test:base_test',
-    'tests/python/pants_test:task_test_base',
+    'tests/python/pants_test/tasks:task_test_base',
   ]
 )
