+++ tests/python/pants_test/backend/jvm/tasks/test_ivy_imports.py
@@ -17,7 +17,7 @@ from pants.backend.jvm.targets.unpacked_jars import UnpackedJars
 from pants.backend.jvm.tasks.ivy_imports import IvyImports
 from pants.base.build_file_aliases import BuildFileAliases
 from pants.util.contextutil import open_zip, temporary_dir
-from pants_test.task_test_base import TaskTestBase
+from pants_test.tasks.task_test_base import TaskTestBase
 
 
 class IvyImportsTest(TaskTestBase):
