+++ tests/python/pants_test/backend/project_info/tasks/BUILD
@@ -24,7 +24,7 @@ python_tests(
     'src/python/pants/backend/python/targets:python',
     'src/python/pants/backend/python:python_requirement',
     'src/python/pants/base:exceptions',
-    'tests/python/pants_test/tasks:base',
+    'tests/python/pants_test/tasks:task_test_base',
   ]
 )
 
@@ -42,7 +42,7 @@ python_tests(
     'src/python/pants/backend/python:plugin',
     'src/python/pants/base:build_file_aliases',
     'src/python/pants/base:exceptions',
-    'tests/python/pants_test/tasks:base',
+    'tests/python/pants_test/tasks:task_test_base',
   ]
 )
 
@@ -59,7 +59,7 @@ python_tests(
     'src/python/pants/backend/project_info/tasks:export',
     'src/python/pants/backend/python:plugin',
     'src/python/pants/base:exceptions',
-    'tests/python/pants_test/tasks:base',
+    'tests/python/pants_test/tasks:task_test_base',
   ]
 )
 
@@ -73,7 +73,7 @@ python_tests(
     'src/python/pants/backend/jvm/targets:java',
     'src/python/pants/backend/project_info/tasks:filedeps',
     'src/python/pants/base:config',
-    'tests/python/pants_test/tasks:base',
+    'tests/python/pants_test/tasks:task_test_base',
   ],
 )
