+++ tests/python/pants_test/backend/project_info/tasks/test_dependencies.py
@@ -13,7 +13,7 @@ from pants.backend.project_info.tasks.dependencies import Dependencies
 from pants.backend.python.python_requirement import PythonRequirement
 from pants.backend.python.targets.python_library import PythonLibrary
 from pants.backend.python.targets.python_requirement_library import PythonRequirementLibrary
-from pants_test.tasks.test_base import ConsoleTaskTestBase
+from pants_test.tasks.task_test_base import ConsoleTaskTestBase
 
 
 class DependenciesEmptyTest(ConsoleTaskTestBase):
