+++ tests/python/pants_test/backend/project_info/tasks/test_depmap.py
@@ -24,7 +24,7 @@ from pants.backend.jvm.targets.scala_library import ScalaLibrary
 from pants.backend.project_info.tasks.depmap import Depmap
 from pants.backend.python.register import build_file_aliases as register_python
 from pants.base.exceptions import TaskError
-from pants_test.tasks.test_base import ConsoleTaskTestBase
+from pants_test.tasks.task_test_base import ConsoleTaskTestBase
 
 
 class BaseDepmapTest(ConsoleTaskTestBase):
