+++ tests/python/pants_test/backend/project_info/tasks/test_filedeps.py
@@ -14,7 +14,7 @@ from pants.backend.jvm.register import build_file_aliases as register_jvm
 from pants.backend.jvm.targets.java_library import JavaLibrary
 from pants.backend.project_info.tasks.filedeps import FileDeps
 from pants.base.config import Config
-from pants_test.tasks.test_base import ConsoleTaskTestBase
+from pants_test.tasks.task_test_base import ConsoleTaskTestBase
 
 
 class FileDepsTest(ConsoleTaskTestBase):
