+++ tests/python/pants_test/tasks/test_builddict.py
@@ -13,7 +13,7 @@ from pants.backend.core.tasks import builddictionary, reflect
 from pants.backend.jvm.register import build_file_aliases as register_jvm
 from pants.backend.python.register import build_file_aliases as register_python
 from pants_test.base_test import BaseTest
-from pants_test.task_test_base import TaskTestBase
+from pants_test.tasks.task_test_base import TaskTestBase
 
 
 OUTDIR = '/tmp/dist'
