+++ tests/python/pants_test/tasks/test_check_published_deps.py
@@ -16,7 +16,7 @@ from pants.backend.jvm.targets.jar_library import JarLibrary
 from pants.backend.jvm.targets.java_library import JavaLibrary
 from pants.backend.jvm.tasks.check_published_deps import CheckPublishedDeps
 from pants.base.build_file_aliases import BuildFileAliases
-from pants_test.tasks.test_base import ConsoleTaskTestBase
+from pants_test.tasks.task_test_base import ConsoleTaskTestBase
 
 
 class CheckPublishedDepsTest(ConsoleTaskTestBase):
