+++ tests/python/pants_test/tasks/test_dependees.py
@@ -19,7 +19,7 @@ from pants.backend.python.targets.python_tests import PythonTests
 from pants.base.build_file_aliases import BuildFileAliases
 from pants.base.exceptions import TaskError
 from pants.base.source_root import SourceRoot
-from pants_test.tasks.test_base import ConsoleTaskTestBase
+from pants_test.tasks.task_test_base import ConsoleTaskTestBase
 
 
 class BaseReverseDepmapTest(ConsoleTaskTestBase):
