+++ tests/python/pants_test/tasks/test_filemap.py
@@ -12,7 +12,7 @@ from pants.backend.core.tasks.filemap import Filemap
 from pants.backend.python.targets.python_library import PythonLibrary
 from pants.base.build_file_aliases import BuildFileAliases
 from pants.base.source_root import SourceRoot
-from pants_test.tasks.test_base import ConsoleTaskTestBase
+from pants_test.tasks.task_test_base import ConsoleTaskTestBase
 
 
 class FilemapTest(ConsoleTaskTestBase):
