+++ tests/python/pants_test/tasks/test_jar_publish.py
@@ -23,7 +23,7 @@ from pants.scm.scm import Scm
 from pants.util.contextutil import temporary_dir
 from pants.util.dirutil import safe_mkdir, safe_walk
 from pants_test.backend.jvm.tasks.jvm_compile.utils import set_compile_strategies
-from pants_test.task_test_base import TaskTestBase
+from pants_test.tasks.task_test_base import TaskTestBase
 
 
 class JarPublishTest(TaskTestBase):
