+++ tests/python/pants_test/tasks/test_jvm_run.py
@@ -10,7 +10,7 @@ import os
 from pants.backend.jvm.targets.jvm_binary import JvmBinary
 from pants.backend.jvm.tasks.jvm_run import JvmRun
 from pants.util.contextutil import pushd, temporary_dir
-from pants_test.task_test_base import TaskTestBase
+from pants_test.tasks.task_test_base import TaskTestBase
 
 
 class JvmRunTest(TaskTestBase):
