+++ tests/python/pants_test/tasks/test_roots.py
@@ -12,7 +12,7 @@ from pants.backend.core.tasks.roots import ListRoots
 from pants.base.build_environment import get_buildroot
 from pants.base.source_root import SourceRoot
 from pants.base.target import Target
-from pants_test.tasks.test_base import ConsoleTaskTestBase
+from pants_test.tasks.task_test_base import ConsoleTaskTestBase
 
 
 @contextmanager
